+++ src/gui/widgets/view_transform_widget.hpp
@@ -22,6 +22,7 @@ public:
 public slots:
     void set_zoom(qreal percent);
     void set_angle(qreal radians);
+    void set_flip(bool flipped);
 
 signals:
     void zoom_in();
@@ -29,6 +30,7 @@ signals:
     void zoom_changed(qreal percent);
     void angle_changed(qreal radians);
     void view_fit();
+    void flip_view();
 
 private slots:
     void fuckyoumoc_on_zoom_changed(qreal percent);